fix errors, reverse back console to page

+++ Dziekanat/StudentsList.cpp
@@ -5,6 +5,8 @@
 StudentsList::StudentsList()
 {
 	this->startIndex = 16000;
+
+	
 }
 
 void StudentsList::addStudent(Student &s)
@@ -20,7 +22,7 @@ void StudentsList::removeStudent()
 void StudentsList::showAllStudents()
 {
 	std::cout << "#################################" << std::endl;
-	for (int i = 0; i < this->students.size(); i++)
+	for (int i = 0; i < (int)this->students.size(); i++)
 	{
 		this->students[i].show();
 	}
@@ -28,8 +30,7 @@ void StudentsList::showAllStudents()
 
 void StudentsList::showStudentbyIndex(int index)
 {
-
-	for (int i = 0; i < this->students.size(); i++)
+	for (int i = 0; i < (int)this->students.size(); i++)
 	{
 		if (this->students[i].getIndex() == index)
 			this->students[i].show();
